// generated-client/api.ts
export interface CreateTask {
    /**
     * Content
     */
    'text': string;
}
/**
 * 
 */
export interface Task {
    /**
     * Identification
     */
    'id': string;
    /**
     * Content
     */
    'text': string;
    /**
     * Status
     */
    'completed': boolean;
    /**
     * Date when task was created (Timestamp)
     */
    'createdDate': number;
    /**
     * Date when task was completed (Timestamp)
     */
    'completedDate'?: number;
}
/**
 * 
 */
export interface UpdateTask {
    /**
     * Content
     */
    'text': string;
}
// ...
export class TasksApi extends BaseAPI {
    /**
     * 
     * @summary Returns all completed tasks
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public tasksCompletedGet(options?: RawAxiosRequestConfig) {
        return TasksApiFp(this.configuration).tasksCompletedGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Returns all tasks. Slow service, around 3 seconds
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public tasksGet(options?: RawAxiosRequestConfig) {
        return TasksApiFp(this.configuration).tasksGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Completes given task, then returns modified task
     * @param {string} id ID of task
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public tasksIdCompletePost(id: string, options?: RawAxiosRequestConfig) {
        return TasksApiFp(this.configuration).tasksIdCompletePost(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Deletes given task
     * @param {string} id ID of task
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public tasksIdDelete(id: string, options?: RawAxiosRequestConfig) {
        return TasksApiFp(this.configuration).tasksIdDelete(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Incompletes given task, then returns modified task
     * @param {string} id ID of task
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public tasksIdIncompletePost(id: string, options?: RawAxiosRequestConfig) {
        return TasksApiFp(this.configuration).tasksIdIncompletePost(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Updates text of given task, then returns modified task
     * @param {string} id ID of task
     * @param {UpdateTask} updateTask text
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public tasksIdPost(id: string, updateTask: UpdateTask, options?: RawAxiosRequestConfig) {
        return TasksApiFp(this.configuration).tasksIdPost(id, updateTask, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Creates task with given text, then returns created task
     * @param {CreateTask} createTask text
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public tasksPost(createTask: CreateTask, options?: RawAxiosRequestConfig) {
        return TasksApiFp(this.configuration).tasksPost(createTask, options).then((request) => request(this.axios, this.basePath));
    }
}